refactor(ui): move user provider HOC
