+++ app/portainer/hooks/useUser.tsx
@@ -8,7 +8,6 @@ import {
   useState,
   useMemo,
   PropsWithChildren,
-  ComponentType,
 } from 'react';
 
 import { isAdmin } from '@/portainer/users/user.helpers';
@@ -177,25 +176,4 @@ export function UserProvider({ children }: UserProviderProps) {
     const user = await getUser(id);
     setUser(user);
   }
-}
-
-export function withCurrentUser<T>(
-  WrappedComponent: ComponentType<T>
-): ComponentType<T> {
-  // Try to create a nice displayName for React Dev Tools.
-  const displayName =
-    WrappedComponent.displayName || WrappedComponent.name || 'Component';
-
-  function WrapperComponent(props: T) {
-    return (
-      <UserProvider>
-        {/* eslint-disable-next-line react/jsx-props-no-spreading */}
-        <WrappedComponent {...props} />
-      </UserProvider>
-    );
-  }
-
-  WrapperComponent.displayName = displayName;
-
-  return WrapperComponent;
 }
