+++ app/portainer/react/views/sidebar.ts
@@ -3,7 +3,7 @@ import angular from 'angular';
 import { AngularSidebarService } from '@/react/sidebar/useSidebarState';
 import { Sidebar } from '@/react/sidebar/Sidebar';
 import { r2a } from '@/react-tools/react2angular';
-import { withCurrentUser } from '@/portainer/hooks/useUser';
+import { withCurrentUser } from '@/react-tools/withCurrentUser';
 import { withReactQuery } from '@/react-tools/withReactQuery';
 import { withUIRouter } from '@/react-tools/withUIRouter';
