+++ app/portainer/settings/edge-compute/EdgeComputeSettingsView.tsx
@@ -1,4 +1,4 @@
-import { withCurrentUser } from '@/portainer/hooks/useUser';
+import { withCurrentUser } from '@/react-tools/withCurrentUser';
 import { r2a } from '@/react-tools/react2angular';
 import { withReactQuery } from '@/react-tools/withReactQuery';
